The path is now printed on the map, improvements on the number of fps and asteroids now appear randomly close enough to paths but not overlapping them or each other

+++ src/Model/Asteroid.java
@@ -1,17 +1,17 @@
 package Model;
 
 public class Asteroid {
-    private int pos_x, pos_y;
+    private double pos_x, pos_y;
     private boolean occupied;
 
-    public Asteroid(int pos_x, int pos_y){
+    public Asteroid(double pos_x, double pos_y){
         this.pos_x = pos_x;
         this.pos_y = pos_y;
         occupied = false;
     }
 
-    public int get_pos_x(){return pos_x;}
-    public int get_pos_y(){return pos_y;}
+    public double get_pos_x(){return pos_x;}
+    public double get_pos_y(){return pos_y;}
 
     public boolean is_occupied() {return occupied;}
     public void occupy(Tower tower){occupied = true;}
